+++ Assets/GridSort/Resources/cginc/GridConfig.cginc
@@ -1,12 +1,12 @@
 #define BITONIC_BLOCK_SIZE 512
 #define TRANSPOSE_BLOCK_SIZE 16
 
-int _BoundaryXMin;
-int _BoundaryXMax;
-int _BoundaryYMin;
-int _BoundaryYMax;
-int _BoundaryZMin;
-int _BoundaryZMax;
+float _BoundaryXMin;
+float _BoundaryXMax;
+float _BoundaryYMin;
+float _BoundaryYMax;
+float _BoundaryZMin;
+float _BoundaryZMax;
 int _GridNumX;
 int _GridNumY;
 int _GridNumZ;
